Update issue 3969 Labels: ReadyForRelease
1. Use the same detection for regexes as the FF version. So we won't miss any of them anymore
2. new RegExp(X) was used twice on complicated regexes. Therefore it got a '/' inserted twice on the begin and end, making it an invalid regex that would never match

Tested on both seti.ee and www.pajacyk.pl and now they are detected and blocked...

p.s.: you can get rid of this:
" && !rule.match(/^\/[a-zA-Z0-9\-\/]+\/$/)"
However, this will make /images/ads/ and that kind of filters be detected as regexes too. But it should still work. I don't know what is the fastest...