Update issue 5160 Status: Started
This should prevent new issue 5115 like issues...

And I assume I know another one of that type:
adreport page -> second question -> select a language of which you are not subscribed to the list for that language -> the message telling you that you have to subscribe to list X is missing the list name.
John, could you:
1. check if I am correct about the possible bug
if YES> Does this proposed fix fix it?
if NO> If you temporary undo your change from issue 5115 and only apply this change, is issue 5115 still fixed after that?
If any of the above is true, please mark it as ReadyForRelease