Update issue 6360 Status: NeedsReview Owner: fam.lam@live.nl CC: adblockforchrome
Address some comments. Also fix a small issue on a plain svg page where document.body doesn't exist. 


>At the end of adblock_start.js: so xml/svg pages cause document.documentElement to be something other than HTMLElement?  That's handy.
Yep. (At least the ones I checked on)

>adblock_start_common.js:32 or so: you call  [...]
Actually, line 32 is a "return;" line which is triggered with exactly that reason, so that check later in the function isn't needed.

>-  css_chunk.style.display = "none !important";
 +  css_chunk.style.setProperty("display", "none", "important");
theoretically, no (both work in Chrome 17), but for some unknown reason Chrome 18 isn't reliable on using el.style.property = "value".

>you might as well use elems.forEach(function(el) {...}) instead of a for loop
elems isn't an array (so forEach is not defined)... (don't ask me what it is, it's no array, but it isn't a normal object either...)