Update Issue 4530 Thanks! Your changes will appear in the next version.
Ideally, it would be nice if all the strings were translated, unless it's impossible to word it that way in your language. Also, there has been a small misunderstanding about the strings blocked_n_in_total and blocked_n_on_this_page, because while it used to say "Do not translate" for the placeholder string, you had to actually copy the string without changing it. So if you could at least update those two, it would be really nice! Thanks again!